...lol
well, if `cargo c` per task on the powerset of 20+ features that wiwi has, if it took "only" 1 second... according to my crude maths we'd be waiting for more than a month (and the longest months too, 31 days), so that's not at all feasible XD

But also, do I really need that? I don't think so. The features are designed to be additive anyways, so its more about enabling one feature and seeing if it enables all the dependency features/crates it needs. Also, do we need matrix? I don't really think so, this is more "check if it compiles", running the test with matrix is good enough I think