this Initialiser/InitState API isn't gonna work (well enough) right now
if/once negative impls (ie. `where T: !Trait`) become a thing, I could use this to satisfy the orphan rules more easily without ex. concrete match on `Init` and `Uninit` struct for a type parameter. For now, setting a value will have to be "eager" in that they are converted before being stored into the builder struct. Unfortunate, but only hypothetically, cause in reality, it won't really actually make a difference in how code runs and performance and etc.